Few updates
* Move few common EC key related functions to separate header file, use
  it in older and in newer implementation
* Move test of MAX_CURVE_NAME_LEN to separate openssl-related test file,
  compile/run that test only with OpenSSL 3
* Remove unneeded comments
